API-12: security: only the person who create the question can archive the same question

+++ app/Http/Controllers/Question/ArchiveController.php
@@ -10,6 +10,7 @@ class ArchiveController extends Controller
 {
     public function __invoke(Question $question): Response
     {
+        $this->authorize('archive', $question);
         $question->delete();
 
         return response()->noContent();
